Add files via upload
RoboSpeaker is a project that lets you create a voice interface that sounds like a human. You can use RoboSpeaker for many things, such as making your own virtual assistant, chatbot, or learning tool. You can also make RoboSpeaker listen and talk back to you, so you can have a conversation with it. RoboSpeaker is made with Python, a programming language that has many libraries for text-to-speech. This means you can use Python to turn text into voice and vice versa. RoboSpeaker is a fun and easy way to learn about voice technology and Python.